% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_sync}
\alias{data_sync}
\title{One-way sync of data from Synapse to local}
\usage{
data_sync()
}
\description{
One-way sync of data from Synapse to local
}
\details{
Looks at all datasets in project_config.yml and pulls them if they aren't recent. To push discovered and derived data back to Synapse, use \code{\link{data_publish}}.
}
