% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_view}
\alias{data_view}
\title{View a dataset on Synapse}
\usage{
data_view(id)
}
\arguments{
\item{id}{Synapse ID of the data file}
}
\description{
View a dataset on Synapse
}
\note{
To see a list of Synapse IDs for registered data files, use \code{\link{data_list}}.
}
