% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{synapse_login}
\alias{synapse_login}
\title{Log in to Synapse}
\usage{
synapse_login(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[synapser]{synLogin}}}
}
\description{
Log in to Synapse
}
