% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{use_apache_license}
\alias{use_apache_license}
\title{Add Apache 2.0 license to a Ki-related R package}
\usage{
use_apache_license(pkg = ".",
  copyright_holder = "Bill and Melinda Gates Foundation")
}
\arguments{
\item{pkg}{package description, can be path or package name. See \code{\link[devtools]{as.package}} for more information.}

\item{copyright_holder}{The copyright holder for the package.}
}
\description{
Add Apache 2.0 license to a Ki-related R package
}
