% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{use_lintr_test}
\alias{use_lintr_test}
\title{Add a lintr checking unit test to your R package}
\usage{
use_lintr_test(pkg = ".")
}
\arguments{
\item{pkg}{package description, can be path or package name. See \code{\link[devtools]{as.package}} for more information.}
}
\description{
Add a lintr checking unit test to your R package
}
