% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_analysis.R
\name{create_analysis}
\alias{create_analysis}
\title{Scaffolding to create a Ki analysis project}
\usage{
create_analysis(path = ".")
}
\arguments{
\item{path}{path to directory where project will be set up}
}
\description{
Scaffolding to create a Ki analysis project
}
