% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_publish}
\alias{data_publish}
\title{Publish a discovered or derived dataset to Synapse}
\usage{
data_publish(obj, name = NULL, type = c("discovered", "derived"),
  file_type = c("rds", "csv", "txt"), desc = NULL, used = NULL,
  executed = NULL, activity = NULL, activity_name = NULL,
  activity_desc = NULL)
}
\arguments{
\item{obj}{object to publish or a path to the file in your project's data/discovered or data/derived folder}

\item{name}{name of the dataset being published}

\item{type}{either "derived" or "discovered". If "discovered", it will go in data/discovered. If "derived", it will go in data/derived.}

\item{file_type}{one of "rds", or "csv"}

\item{desc}{description of the dataset}

\item{used}{(optional) The Entity, Synapse ID, or URL used to create the object (can also be a list of these)}

\item{executed}{(optional) The Entity, Synapse ID, or URL representing code executed to create the object (can also be a list of these)}

\item{activity}{(optional) Activity object specifying the user's provenance}

\item{activity_name}{(optional) Activity name to be used in conjunction with *used* and *executed*.}

\item{activity_desc}{(optional) Activity description to be used in conjunction with *used* and *executed*.}
}
\description{
Publish a discovered or derived dataset to Synapse
}
