% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_use}
\alias{data_use}
\title{Register a Synapse dataset with your analysis and download a local copy}
\usage{
data_use(id, type = c("core", "discovered", "derived"))
}
\arguments{
\item{id}{Synapse ID of dataset}

\item{type}{one of "core", "discovered", or "derived"}
}
\description{
Register a Synapse dataset with your analysis and download a local copy
}
